refactor: Split types.rs into modular components
Split monolithic types.rs into separate modules for better organization and maintainability. Each component of the MCP schema now lives in its own file under src/types/:

- base.rs: Core JSON-RPC and base MCP types
- common.rs: Shared types and utilities
- initialization.rs: Protocol initialization types
- resources.rs: Resource management types
- prompts.rs: Prompt handling types
- tools.rs: Tool-related types
- etc.

This reorganization improves:
- Code organization and findability
- Module-level documentation
- Testing isolation
- Maintainability of individual components

Version bump: 0.1.2 -> 0.1.3
