+++ WORK_NOTES.md
@@ -1,0 +1,87 @@
+# MCP Schema Rustプロジェクト 作業メモ
+
+## 完了した作業
+
+### 1. プロジェクト構造の整理
+- 元の`types.rs`を複数のモジュールに分割
+- `src/types/`ディレクトリを作成
+- 各機能ごとのモジュールファイルを作成
+
+### 2. 作成したファイル
+```
+src/
+├── lib.rs
+└── types/
+    ├── mod.rs
+    ├── base.rs
+    ├── cancellation.rs
+    ├── client.rs
+    ├── client_completion.rs
+    ├── common.rs
+    ├── initialization.rs
+    ├── logging.rs
+    ├── progress.rs
+    ├── prompts.rs
+    ├── resources.rs
+    ├── roots.rs
+    ├── sampling.rs
+    ├── server.rs
+    └── tools.rs
+```
+
+### 3. 基本設定ファイルの更新
+- `Cargo.toml`の更新（依存関係、メタデータの追加）
+- `lib.rs`の更新（ドキュメント、再エクスポート、基本テストの追加）
+
+### 4. ドキュメントとテストの追加（一部）
+- `base.rs`にドキュメントとテストを追加
+  - 型の説明
+  - 使用例
+  - シリアライゼーションテスト
+
+## これからやるべき作業
+
+### 1. 残りのモジュールのドキュメント化
+以下のファイルにドキュメントとテストを追加する必要があります：
+- [ ] common.rs
+- [ ] initialization.rs
+- [ ] resources.rs
+- [ ] prompts.rs
+- [ ] tools.rs
+- [ ] logging.rs
+- [ ] sampling.rs
+- [ ] client.rs
+- [ ] server.rs
+- [ ] cancellation.rs
+- [ ] progress.rs
+- [ ] client_completion.rs
+- [ ] roots.rs
+
+### 2. 統合テストの追加
+- [ ] 実際のユースケースに基づくテストシナリオの作成
+- [ ] エラーハンドリングのテスト
+- [ ] シリアライゼーション/デシリアライゼーションの包括的なテスト
+
+### 3. CI/CD設定
+- [ ] GitHub Actionsの設定
+- [ ] テスト自動化
+- [ ] コードカバレッジレポート
+- [ ] ドキュメント生成
+
+### 4. ドキュメント
+- [ ] READMEの充実化
+- [ ] API文書の生成
+- [ ] 使用例の追加
+
+### 5. その他
+- [ ] エラー型の改善
+- [ ] ビルダーパターンの導入検討
+- [ ] より厳密なバリデーションの追加
+- [ ] パフォーマンス最適化
+
+## 注意点
+1. テストカバレッジを高く保つ
+2. ドキュメントは実際の使用例を含める
+3. APIの使いやすさを重視
+4. 型安全性を確保
+5. エラーメッセージを明確に
