+++ src/types/cancellation.rs
@@ -1,0 +1,11 @@
+use serde::{Deserialize, Serialize};
+use super::base::RequestId;
+
+/// Parameters for a cancelled-notification
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct CancelledNotificationParams {
+    pub request_id: RequestId,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub reason: Option<String>,
+}
