+++ src/types/client_completion.rs
@@ -1,0 +1,56 @@
+use serde::{Deserialize, Serialize};
+use serde_json::Value;
+use std::collections::HashMap;
+
+/// A reference to either a resource or a prompt
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase", tag = "type")]
+pub enum ReferenceType {
+    #[serde(rename = "ref/resource")]
+    Resource { uri: String },
+    #[serde(rename = "ref/prompt")]
+    Prompt { name: String },
+}
+
+/// An argument for completion/complete
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct CompleteArgument {
+    pub name: String,
+    pub value: String,
+    #[serde(flatten)]
+    pub extra: HashMap<String, Value>,
+}
+
+/// Parameters for completion/complete
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct CompleteParams {
+    #[serde(rename = "ref")]
+    pub r#ref: ReferenceType,
+    pub argument: CompleteArgument,
+    #[serde(flatten)]
+    pub extra: HashMap<String, Value>,
+}
+
+/// Data containing possible completions
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct CompletionData {
+    pub values: Vec<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub total: Option<i64>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub has_more: Option<bool>,
+}
+
+/// Result from completion/complete
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct CompleteResult {
+    #[serde(rename = "_meta", skip_serializing_if = "Option::is_none")]
+    pub meta: Option<HashMap<String, Value>>,
+    pub completion: CompletionData,
+    #[serde(flatten)]
+    pub extra: HashMap<String, Value>,
+}
